+++ examples/lib/blocks.jl
@@ -7,10 +7,10 @@
 
 export PID_Controller
 
-@comment """
+"""
 # Blocks
 """
-
+@comment 
 
 """
 Demonstrates the usage of a Continuous.LimPID controller