+++ examples/lib/heat_transfer.jl
@@ -7,9 +7,10 @@
 
 export TwoMasses, Motor
 
-@comment """
+"""
 # Heat transfer
 """
+@comment 
 
 """
 Simple conduction demo