+++ examples/lib/rotational.jl
@@ -8,9 +8,10 @@
 
 export First
 
-@comment """
+"""
 # Rotational
 """
+@comment 
 
 """
 First example: simple drive train