+++ lib/heat_transfer.jl
@@ -4,7 +4,7 @@
 ## Heat transfer models               ##
 ########################################
 
-@comment """
+"""
 # Heat transfer models
 
 Library of 1-dimensional heat transfer with lumped elements
@@ -23,26 +23,31 @@ usually not possible).
 Note, that all temperatures of this package, including initial
 conditions, are given in Kelvin.
 """
+@comment 
 
 
 ########################################
 ## Basic
 ########################################
-@comment """
+"""
 ## Basics
 """
+@comment 
 
 
 """
 Lumped thermal element storing heat
 
 ```julia
-HeatCapacitor(hp::HeatPort, C::Signal)
+HeatCapacitor(hp::HeatPort; C::Signal)
 ```
 
 ### Arguments
 
 * `hp::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `C::Signal` : heat capacity of the element [J/K]
 
 ### Details
@@ -71,7 +76,7 @@ cp of the component material to compute C:
 ```
 
 NOTE: The Modelica Standard Library has an argument Tstart for the
-starting temperature [K]. You really can't used that here as in
+starting temperature [K]. You really can't use that here as in
 Modelica. You need to define the starting temperature at the top level
 for the HeatPort you define.
 
@@ -89,13 +94,16 @@ end
 Lumped thermal element transporting heat without storing it
 
 ```julia
-ThermalConductor(port_a::HeatPort, port_b::HeatPort, G::Signal)
+ThermalConductor(port_a::HeatPort, port_b::HeatPort; G::Signal)
 ```
 
 ### Arguments
 
 * `port_a::HeatPort` : heat port [K]
 * `port_b::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `G::Signal` : Constant thermal conductance of material [W/K]
 
 ### Details
@@ -162,6 +170,9 @@ Convection(port_a::HeatPort, port_b::HeatPort; Gc::Signal)
 
 * `port_a::HeatPort` : heat port [K]
 * `port_b::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `Gc::Signal` : convective thermal conductance [W/K]
 
 ### Details
@@ -240,13 +251,16 @@ end
 """
 
 ```julia
-BodyRadiation(port_a::HeatPort, port_b::HeatPort, Gr::Signal)
+BodyRadiation(port_a::HeatPort, port_b::HeatPort; Gr::Signal)
 ```
 
 ### Arguments
 
 * `port_a::HeatPort` : heat port [K]
 * `port_b::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `Gr::Signal` : net radiation conductance between two surfaces [m2]
 
 
@@ -357,9 +371,10 @@ end
 ## Sources
 ########################################
 
-@comment """
+"""
 ## Sources
 """
+@comment 
 
 
 """
@@ -373,12 +388,15 @@ variable. FixedTemperature and PrescribedTemperature are identical;
 naming is for Modelica compatibility.)
 
 ```julia
-FixedTemperature(port::HeatPort, T::Signal)
+FixedTemperature(port::HeatPort; T::Signal)
 ```
 
 ### Arguments
 
 * `port::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `T::Signal` : temperature at port [K]
 
 """
@@ -403,12 +421,15 @@ variable. FixedTemperature and PrescribedTemperature are identical;
 naming is for Modelica compatibility.)
 
 ```julia
-PrescribedTemperature(port::HeatPort, T::Signal)
+PrescribedTemperature(port::HeatPort; T::Signal)
 ```
 
 ### Arguments
 
 * `port::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `T::Signal` : temperature at port [K]
 
 """
@@ -432,13 +453,15 @@ losses (which are given an reference temperature T_ref).
 variable.)
 
 ```julia
-FixedHeatFlow(port::HeatPort, Q_flow::Signal, T_ref::Signal = 293.15, alpha::Signal = 0.0)
-FixedHeatFlow(port::HeatPort, Q_flow::Signal; T_ref::Signal = 293.15, alpha::Signal = 0.0)
+FixedHeatFlow(port::HeatPort; Q_flow::Signal, T_ref::Signal = 293.15, alpha::Signal = 0.0)
 ```
 
 ### Arguments
 
 * `port::HeatPort` : heat port [K]
+
+### Keyword/Optional Arguments
+
 * `Q_flow::Signal` : heat flow [W]
 
 ### Keyword/Optional Arguments
@@ -471,7 +494,6 @@ losses (which are given an reference temperature T_ref).
 variable.)
 
 ```julia
-PrescribedHeatFlow(port::HeatPort, Q_flow::Signal, T_ref::Signal = 293.15, alpha::Signal = 0.0)
 PrescribedHeatFlow(port::HeatPort, Q_flow::Signal; T_ref::Signal = 293.15, alpha::Signal = 0.0)
 ```